Refactor the React files
