+++ templates/tournament/index.html.twig
@@ -4,25 +4,27 @@
 
 {% block body %}
     <div class="row p-3 justify-content-center">
-        <div class="mx-auto">
+        <div class="col-md-10 col-lg-8">
             <div class="col-12">
                 <img class="img-fluid" src="{{ asset('images/tournament-index.jpeg') }}" />
             </div>
-            <div>
-                <h1 class="mb-5">Welcome to the tournament!</h1>        
+            {% if isAllowed is same as(true) %}
+                <div class="row my-3 justify-content-center">
+                    <a href="{{ path('tournament_battle') }}"><button class="btn btn-primary">{{ buttonMessage }}</button></a>
+                </div>
+            {% endif %}
+            <div class="mx-auto">
+                <h1 class="my-3">Welcome to the tournament!</h1>        
 
-                <p>You need at least 3 pokemons in shape to participate at the tournament.</p>
-                <p>
-                    This is a battle between 2 trainers with <span class="text-success font-weight-bold">3 pokemons</span> each.<br><br>
-                    You're also allowed to use <span class="text-success font-weight-bold">3 health potions</span> maximum during a battle.<br>
-                    So don't forget to bring some with you!
+                <p>You need at least 3 pokemons in shape to participate at the tournament. This is a battle between 2 trainers with <span class="text-success font-weight-bold">3 pokemons</span> each.<br><br>
+                    You're also allowed to use <span class="text-success font-weight-bold">3 health potions</span> maximum during a battle. So don't forget to bring some with you!
                 </p>
                 <p>Overcome 3 battles in a row to win <span class="text-warning font-weight-bold">the tournament</span>!</p>
             </div>
         </div>
     </div>
     {% if isAllowed is same as(true) %}
-        <div class="row my-3 justify-content-center">
+        <div class="row my-2 justify-content-center">
             <a href="{{ path('tournament_battle') }}"><button class="btn btn-primary">{{ buttonMessage }}</button></a>
         </div>
     {% endif %}
